threshold limit added, gauge default parameters moved to constants/explorer
