+++ dashboards-observability/common/constants/explorer.ts
@@ -132,3 +132,18 @@ export const ROW_DENSITIES = [
 ];
 
 export const HEADER_HEIGHT = 35;
+
+// gauge chart default parameters
+export interface DefaultGaugeChartParametersProps {
+  GaugeTitleSize: number,
+  DisplayDefaultGauges: number,
+  OrientationDefault: string,
+  TickLength: number, 
+};
+
+export const DefaultGaugeChartParameters: DefaultGaugeChartParametersProps = {
+  GaugeTitleSize: 14,
+  DisplayDefaultGauges: 1,
+  OrientationDefault: 'h',
+  TickLength: 5,
+}
