+++ dashboards-observability/public/components/event_analytics/explorer/visualizations/config_panel/config_panes/config_controls/config_panel_option_gauge.tsx
@@ -5,10 +5,8 @@
 
 import React from 'react';
 import { EuiFormRow, EuiFieldNumber } from '@elastic/eui';
-import {
-  visChartTypes,
-  DefaultGaugeChartParameters,
-} from '../../../../../../../../common/constants/shared';
+import { visChartTypes } from '../../../../../../../../common/constants/shared';
+import { DefaultGaugeChartParameters } from '../../../../../../../../common/constants/explorer';
 
 export const ConfigPanelOptionGauge = ({
   visualizations,
