+++ dashboards-observability/public/components/event_analytics/explorer/visualizations/config_panel/config_panes/config_controls/config_panel_options.tsx
@@ -10,12 +10,8 @@ import {
   EuiFormRow,
   EuiTextArea,
   EuiAccordion,
-  EuiFieldNumber,
 } from '@elastic/eui';
-import {
-  visChartTypes,
-  DefaultGaugeChartParameters,
-} from '../../../../../../../../common/constants/shared';
+import { visChartTypes } from '../../../../../../../../common/constants/shared';
 import { ConfigPanelOptionGauge } from './config_panel_option_gauge';
 const helpText = 'Name your visualization.';
