+++ dashboards-observability/public/components/event_analytics/explorer/visualizations/config_panel/config_panes/config_controls/config_thresholds.tsx
@@ -32,6 +32,7 @@ export const ConfigThresholds = ({
   vizState = [],
   handleConfigChange,
   sectionName = 'Thresholds',
+  props,
 }: any) => {
   const addButtonText = '+ Add threshold';
   const getThresholdUnit = () => {
@@ -87,6 +88,9 @@ export const ConfigThresholds = ({
         fullWidth
         size="s"
         onClick={handleAddThreshold}
+        {...(props?.maxLimit && {
+          isDisabled: !isEmpty(vizState) && vizState.length === props?.maxLimit,
+        })}
       >
         {addButtonText}
       </EuiButton>
