+++ dashboards-observability/public/components/event_analytics/explorer/visualizations/config_panel/config_panes/default_vis_editor.tsx
@@ -28,6 +28,7 @@ export const VizDataPanel = ({ visualizations, onConfigChange, vizState = {}, ta
           vizState={vizState[section.mapTo] || section.defaultState || {}}
           sectionName={section.name}
           sectionId={section.id}
+          props={section.props || {}}
         />
       </EuiFormRow>
     );
